장소 검색 API distance(반경) 반영  (#205)
* 검색 화면 반경 1km 내에 위치한 장소만 마커가 찍히도록 로직 변경

거리, 위도, 경도 API 규격에 맞도록 Int -> Double 로 타입 변경

* TokenAuthenticator 내에서 token 을 제거하는 로직 주석처리

* 검색된 장소만 맵에 마커가 찍히도록 로직수정

* 검색 화면 반경 거리 수정

1km -> 2.5km 네이버 지도 참고

+++ data/src/main/kotlin/us/wedemy/eggeum/android/data/datasource/place/remote/PlaceRemoteDataSource.kt
@@ -16,10 +16,10 @@ public interface PlaceRemoteDataSource {
   public suspend fun getPlace(placeId: Long): PlaceResponse?
 
   public fun getPlaceList(
-    distance: Int?,
+    distance: Double?,
     endDate: String?,
-    latitude: Int?,
-    longitude: Int?,
+    latitude: Double?,
+    longitude: Double?,
     page: Int?,
     search: String?,
     size: Int?,
